+++ src/Additions/Additions.php
@@ -106,6 +106,9 @@ class Additions {
 					break;
 			}
 
+			$addition['PrimaryBranch'] = ! empty( $repo['primary_branch'] ) ? $repo['primary_branch'] : 'master';
+			$addition['ReleaseAsset']  = isset( $repo['release_asset'] ) ? 'true' : null;
+
 			$this->add_to_git_updater[ $repo['slug'] ] = array_merge( $additions[ $repo['slug'] ], $addition );
 		}
 	}