+++ CHANGES.md
@@ -1,5 +1,9 @@
 #### [unreleased]
 
+#### 5.4.0 / 2021-08-03
+* add settings for `Primary Branch` and `Release Asset`
+* only use `esc_attr_e` for translating strings
+
 #### 5.3.0 / 2021-07-05
 * update Freemius for multisite
 * remove Freemius from the autoloader